requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
PyYAML==6.0.1
protobuf>=4.0.0
gtfs-realtime-bindings>=1.0.0

